[widgets]: Improve centering by inserting padder spans
We used to mutate the 1st span to pad, however that was not very
reliable, as there might be any number of spans in the line. Instead, we
now add a padding span to index 0, and one to index -1.

We also make a special exception for content==" ", since it's more
efficient to return a single span full of padding than 3 space-filled
ones.