[text]: Add rudimentary routing functionality
This currently looks for the first hyperlink markup tag within its
content, and on press (regardless of where) will try to route wherever
its linked to.

In the future (once proper wrapping support is added) this will detect
which hyperlink tag (if any) was clicked, and not route regardless of
position.