[widgets]: Implement new container arrange algorithm
This is mostly an evolution of the current one, doubling down on its
semantics and questioning them when necessary.

`positioning` and its associated enum are now labeled `anchor`, and have
new values + meanings:

anchor == ...

- Anchor.NONE: position as parent arrangement requires
- Anchor.PARENT: position as parent position + self offset
- Anchor.SCREEN: position as terminal origin + self offset

In addition, gaps are now simplified:

- Only None and integer gaps are now permitted
- gap == None means the gap will be calculated based on the available
  space (think CSS flex justify-between)
- gap == {int} will arrange widgets into a stack with exact gaps of the
  given size between them

The removal of fractional gaps is mostly due to their confusing
semantics - if I set a gap to be 0.3, what do you expect? Should it be
0.3 * parent width, or 0.3 * maximum gap? I'd argue the latter, but it's
not too intuitive. It also introduces a lot of complexity that we
frankly don't need.

Widgets are also no longer drawn if they are outside of their parent's
viewport. In the near future we will also be able to clip widgets to
only display parts that are in the viewport, but that's for another day.

Toodles.